Move filter and sort handlers directly into useViews
